Move `sc_network::protocol::sync` to `sc_network_sync` with submodules, move message data structures around accordingly
