+++ client/network/src/protocol/message.rs
@@ -20,19 +20,13 @@
 //! payload.
 
 pub use self::generic::{
-	BlockAnnounce, FromBlock, RemoteCallRequest, RemoteChangesRequest, RemoteChangesResponse,
-	RemoteHeaderRequest, RemoteHeaderResponse, RemoteReadChildRequest, RemoteReadRequest, Roles,
+	RemoteCallRequest, RemoteChangesRequest, RemoteChangesResponse, RemoteHeaderRequest,
+	RemoteHeaderResponse, RemoteReadChildRequest, RemoteReadRequest, Roles,
 };
-use bitflags::bitflags;
-use codec::{Decode, Encode, Error, Input, Output};
+use codec::{Decode, Encode};
 use sc_client_api::StorageProof;
-use sp_runtime::{
-	traits::{Block as BlockT, Header as HeaderT},
-	ConsensusEngineId,
-};
-
-/// A unique ID of a request.
-pub type RequestId = u64;
+use sc_network_common::message::RequestId;
+use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
 
 /// Type alias for using the message type using block type parameters.
 pub type Message<B> = generic::Message<
@@ -42,81 +36,8 @@ pub type Message<B> = generic::Message<
 	<B as BlockT>::Extrinsic,
 >;
 
-/// Type alias for using the block request type using block type parameters.
-pub type BlockRequest<B> =
-	generic::BlockRequest<<B as BlockT>::Hash, <<B as BlockT>::Header as HeaderT>::Number>;
-
-/// Type alias for using the BlockResponse type using block type parameters.
-pub type BlockResponse<B> =
-	generic::BlockResponse<<B as BlockT>::Header, <B as BlockT>::Hash, <B as BlockT>::Extrinsic>;
-
 /// A set of transactions.
 pub type Transactions<E> = Vec<E>;
-
-// Bits of block data and associated artifacts to request.
-bitflags! {
-	/// Node roles bitmask.
-	pub struct BlockAttributes: u8 {
-		/// Include block header.
-		const HEADER = 0b00000001;
-		/// Include block body.
-		const BODY = 0b00000010;
-		/// Include block receipt.
-		const RECEIPT = 0b00000100;
-		/// Include block message queue.
-		const MESSAGE_QUEUE = 0b00001000;
-		/// Include a justification for the block.
-		const JUSTIFICATION = 0b00010000;
-		/// Include indexed transactions for a block.
-		const INDEXED_BODY = 0b00100000;
-	}
-}
-
-impl BlockAttributes {
-	/// Encodes attributes as big endian u32, compatible with SCALE-encoding (i.e the
-	/// significant byte has zero index).
-	pub fn to_be_u32(&self) -> u32 {
-		u32::from_be_bytes([self.bits(), 0, 0, 0])
-	}
-
-	/// Decodes attributes, encoded with the `encode_to_be_u32()` call.
-	pub fn from_be_u32(encoded: u32) -> Result<Self, Error> {
-		Self::from_bits(encoded.to_be_bytes()[0])
-			.ok_or_else(|| Error::from("Invalid BlockAttributes"))
-	}
-}
-
-impl Encode for BlockAttributes {
-	fn encode_to<T: Output + ?Sized>(&self, dest: &mut T) {
-		dest.push_byte(self.bits())
-	}
-}
-
-impl codec::EncodeLike for BlockAttributes {}
-
-impl Decode for BlockAttributes {
-	fn decode<I: Input>(input: &mut I) -> Result<Self, Error> {
-		Self::from_bits(input.read_byte()?).ok_or_else(|| Error::from("Invalid bytes"))
-	}
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Encode, Decode)]
-/// Block enumeration direction.
-pub enum Direction {
-	/// Enumerate in ascending order (from child to parent).
-	Ascending = 0,
-	/// Enumerate in descending order (from parent to canonical child).
-	Descending = 1,
-}
-
-/// Block state in the chain.
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Encode, Decode)]
-pub enum BlockState {
-	/// Block is not part of the best chain.
-	Normal,
-	/// Latest best block.
-	Best,
-}
 
 /// Remote call response.
 #[derive(Debug, PartialEq, Eq, Clone, Encode, Decode)]
@@ -136,35 +57,18 @@ pub struct RemoteReadResponse {
 	pub proof: StorageProof,
 }
 
-/// Announcement summary used for debug logging.
-#[derive(Debug)]
-pub struct AnnouncementSummary<H: HeaderT> {
-	pub block_hash: H::Hash,
-	pub number: H::Number,
-	pub parent_hash: H::Hash,
-	pub state: Option<BlockState>,
-}
-
-impl<H: HeaderT> BlockAnnounce<H> {
-	pub fn summary(&self) -> AnnouncementSummary<H> {
-		AnnouncementSummary {
-			block_hash: self.header.hash(),
-			number: *self.header.number(),
-			parent_hash: self.header.parent_hash().clone(),
-			state: self.state,
-		}
-	}
-}
-
 /// Generic types.
 pub mod generic {
-	use super::{
-		BlockAttributes, BlockState, ConsensusEngineId, Direction, RemoteCallResponse,
-		RemoteReadResponse, RequestId, StorageProof, Transactions,
-	};
+	use super::{RemoteCallResponse, RemoteReadResponse, Transactions};
 	use bitflags::bitflags;
 	use codec::{Decode, Encode, Input, Output};
-	use sc_network_sync::message::generic::BlockData;
+	use sc_client_api::StorageProof;
+	use sc_network_common::message::RequestId;
+	use sc_network_sync::message::{
+		generic::{BlockRequest, BlockResponse},
+		BlockAnnounce,
+	};
+	use sp_runtime::ConsensusEngineId;
 
 	bitflags! {
 		/// Bitmask of the roles that a node fulfills.
@@ -228,15 +132,6 @@ pub mod generic {
 		pub protocol: ConsensusEngineId,
 		/// Message payload.
 		pub data: Vec<u8>,
-	}
-
-	/// Identifies starting point of a block sequence.
-	#[derive(Debug, PartialEq, Eq, Clone, Encode, Decode)]
-	pub enum FromBlock<Hash, Number> {
-		/// Start with given hash.
-		Hash(Hash),
-		/// Start with given block number.
-		Number(Number),
 	}
 
 	/// A network message.
@@ -352,68 +247,6 @@ pub mod generic {
 				genesis_hash,
 				chain_status,
 			})
-		}
-	}
-
-	/// Request block data from a peer.
-	#[derive(Debug, PartialEq, Eq, Clone, Encode, Decode)]
-	pub struct BlockRequest<Hash, Number> {
-		/// Unique request id.
-		pub id: RequestId,
-		/// Bits of block data to request.
-		pub fields: BlockAttributes,
-		/// Start from this block.
-		pub from: FromBlock<Hash, Number>,
-		/// End at this block. An implementation defined maximum is used when unspecified.
-		pub to: Option<Hash>,
-		/// Sequence direction.
-		pub direction: Direction,
-		/// Maximum number of blocks to return. An implementation defined maximum is used when
-		/// unspecified.
-		pub max: Option<u32>,
-	}
-
-	/// Response to `BlockRequest`
-	#[derive(Debug, PartialEq, Eq, Clone, Encode, Decode)]
-	pub struct BlockResponse<Header, Hash, Extrinsic> {
-		/// Id of a request this response was made for.
-		pub id: RequestId,
-		/// Block data for the requested sequence.
-		pub blocks: Vec<BlockData<Header, Hash, Extrinsic>>,
-	}
-
-	/// Announce a new complete relay chain block on the network.
-	#[derive(Debug, PartialEq, Eq, Clone)]
-	pub struct BlockAnnounce<H> {
-		/// New block header.
-		pub header: H,
-		/// Block state. TODO: Remove `Option` and custom encoding when v4 becomes common.
-		pub state: Option<BlockState>,
-		/// Data associated with this block announcement, e.g. a candidate message.
-		pub data: Option<Vec<u8>>,
-	}
-
-	// Custom Encode/Decode impl to maintain backwards compatibility with v3.
-	// This assumes that the packet contains nothing but the announcement message.
-	// TODO: Get rid of it once protocol v4 is common.
-	impl<H: Encode> Encode for BlockAnnounce<H> {
-		fn encode_to<T: Output + ?Sized>(&self, dest: &mut T) {
-			self.header.encode_to(dest);
-			if let Some(state) = &self.state {
-				state.encode_to(dest);
-			}
-			if let Some(data) = &self.data {
-				data.encode_to(dest)
-			}
-		}
-	}
-
-	impl<H: Decode> Decode for BlockAnnounce<H> {
-		fn decode<I: Input>(input: &mut I) -> Result<Self, codec::Error> {
-			let header = H::decode(input)?;
-			let state = BlockState::decode(input).ok();
-			let data = Vec::decode(input).ok();
-			Ok(Self { header, state, data })
 		}
 	}
