+++ client/network/sync/Cargo.toml
@@ -17,9 +17,12 @@ targets = ["x86_64-unknown-linux-gnu"]
 prost-build = "0.9"
 
 [dependencies]
+bitflags = "1.3.2"
 codec = { package = "parity-scale-codec", version = "3.0.0", features = [
     "derive",
 ] }
+either = "1.5.3"
+fork-tree = { version = "3.0.0", path = "../../../utils/fork-tree" }
 futures = "0.3.21"
 libp2p = "0.44.0"
 log = "0.4.16"
@@ -27,9 +30,19 @@ prost = "0.9"
 sc-client-api = { version = "4.0.0-dev", path = "../../api" }
 sc-consensus = { version = "0.10.0-dev", path = "../../consensus/common" }
 sc-network-common = { version = "0.10.0-dev", path = "../common" }
+sc-peerset = { version = "4.0.0-dev", path = "../../peerset" }
 smallvec = "1.8.0"
+sp-arithmetic = { version = "5.0.0", path = "../../../primitives/arithmetic" }
 sp-blockchain = { version = "4.0.0-dev", path = "../../../primitives/blockchain" }
+sp-consensus = { version = "0.10.0-dev", path = "../../../primitives/consensus/common" }
 sp-core = { version = "6.0.0", path = "../../../primitives/core" }
 sp-finality-grandpa = { version = "4.0.0-dev", path = "../../../primitives/finality-grandpa" }
 sp-runtime = { version = "6.0.0", path = "../../../primitives/runtime" }
 thiserror = "1.0"
+
+[dev-dependencies]
+quickcheck = "1.0.3"
+sc-block-builder = { version = "0.10.0-dev", path = "../../block-builder" }
+sp-test-primitives = { version = "2.0.0", path = "../../../primitives/test-primitives" }
+sp-tracing = { version = "5.0.0", path = "../../../primitives/tracing" }
+substrate-test-runtime-client = { version = "2.0.0", path = "../../../test-utils/runtime/client" }
